修复copy异常问题



# message auto-generated for no-merge-commit merge:
!201 merge develop into develop

修复copy异常问题

Created-by: chenyz6
Commit-by: chenxin
Merged-by: yrewzjsx
Description: # Pull Request Template

## Description

同步流不对，导致数据未真正完成copy

## Related Issues

<!-- Replace with actual issue numbers. Example: Closes #123, Related to #456 -->

## Type of Change

- [ ] Bug fix
- [ ] New feature
- [ ] Performance optimization
- [ ] Documentation update
- [ ] Other (please describe):_______________

## Testing

<!-- Describe how the changes were tested. Include:
- Unit/integration test coverage
- Manual validation steps
- Links to CI reports (e.g., CIDA) if available
- Note any pending tests
-->

## Checklist

- [ ] I have performed a self-review of my own code.  
- [ ] I have updated the documentation.  
- [ ] I have added tests to prove my changes are effective.

See merge request: Ascend/memfabric_hybrid!201

+++ src/hybm/csrc/copy_extend/hybm_copy_kernel.cpp
@@ -77,10 +77,9 @@ extern "C" __global__ __aicore__ void hybm_copy_kernel(GM_ADDR dst, GM_ADDR src,
     uint32_t idx = AscendC::GetBlockIdx();
     uint32_t num = AscendC::GetBlockNum();
     uint64_t offset = ((len + SINGLE_COPY_SLICE - 1U) / SINGLE_COPY_SLICE + num - 1U) / num * SINGLE_COPY_SLICE;
-    uint32_t size = min(offset * (idx + 1), len);
+    uint64_t size = min(offset * (idx + 1), len);
     offset = offset * idx;
     size -= offset;
-
     copy_gm2gm(dst + offset, src + offset, 0, COPY_BUF_SIZE, size);
 }
